refactor(lotto) : ErrorMessage enum 클래스 이동
ErrorMessage enum은 view와 관련이 깊은 것 같아 클래스 위치를 변경하였습니다.
